♻️ refactor: redisTemplate 호출부 리팩토링 및 테스트코드 수정(2) 전체 작업 완료
