+++ src/main/java/com/project/cheerha/common/elasticsearch/ElasticsearchClientService.java
@@ -7,13 +7,13 @@ import com.project.cheerha.domain.elasticsearch.entity.JobOpeningDocument;
 import com.project.cheerha.common.exception.data.ElasticsearchQueryException;
 import com.project.cheerha.common.exception.data.DataErrorCode;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Component
+@Repository
 @RequiredArgsConstructor
 public class ElasticsearchClientService {
