+++ src/main/java/com/project/cheerha/domain/auth/service/BlackListService.java
@@ -1,7 +1,7 @@
 package com.project.cheerha.domain.auth.service;
 
 import com.project.cheerha.common.properties.JwtSecurityProperties;
-import com.project.cheerha.common.repository.KeyValueRepository;
+import com.project.cheerha.common.repository.keyvalue.KeyValueRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
