+++ src/main/java/com/project/cheerha/domain/user/service/EmailTokenService.java
@@ -2,7 +2,7 @@ package com.project.cheerha.domain.user.service;
 
 import com.project.cheerha.common.exception.client.BadRequestException;
 import com.project.cheerha.common.exception.client.ClientErrorCode;
-import com.project.cheerha.common.repository.KeyValueRepository;
+import com.project.cheerha.common.repository.keyvalue.KeyValueRepository;
 import com.project.cheerha.common.util.SecureRandomUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
