+++ src/test/java/com/project/cheerha/redis/count/IncrementDailyLimitTest.java
@@ -1,7 +1,7 @@
 package com.project.cheerha.redis.count;
 
 import com.project.cheerha.common.exception.client.BadRequestException;
-import com.project.cheerha.common.repository.KeyValueRepository;
+import com.project.cheerha.common.repository.keyvalue.KeyValueRepository;
 import com.project.cheerha.domain.user.service.CheckDailyEmailCount;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
